renamed folder to demo_light
