+++ frontend/demo_light/secrets.toml.example
